requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

sentence-transformers==2.2.2
faiss-cpu==1.9.0

pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2

google-generativeai==0.3.0
streamlit==1.28.1
requests==2.31.0
